libs/genai/pyproject.toml
[tool.poetry]
name = "genai"
version = "0.0.1"
description = "genai tools and tasks"
authors = ["lucasBritoo <britol599@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
genai-core = { path = "../core/", develop = true }
langchain-openai = "^0.1.3"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
